# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Hossi-911 Hospital API
  version: 1.0.0
paths:
  /login:
    post:
      summary: User login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: {type: string}
        '401':
          description: Invalid credentials
  /patients:
    get:
      summary: Get patients (branch-specific)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: {type: integer}
                    name: {type: string}
                    dob: {type: string}
                    gender: {type: string}
        '403':
          description: Access denied
    post:
      summary: Create a patient
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: {type: string}
                password: {type: string}
                name: {type: string}
                dob: {type: string}
                gender: {type: string}
                id_number: {type: string}
                branch_id: {type: integer}
                allergies: {type: string}
                medical_history: {type: string}
      responses:
        '201':
          description: Patient created
        '403':
          description: Access denied
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT